requirements.txt
python-telegram-bot[job-queue]==21.6
trafilatura==1.12.0
beautifulsoup4==4.12.3
requests==2.32.3
requests-oauthlib==1.3.1
youtube-transcript-api==0.6.2
yt-dlp>=2025.1.15
feedparser==6.0.11
googlenewsdecoder>=0.1.7
Pillow>=10.0.0
